Restructured folders and files
Reorganized files and folders to improve productivity and debugging. Further changes to come that separate functions and components into individual files for testing purposes.
