--- Carbon/App.js
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import MainContainer from './navigation/screens/Main/MainContainer';
-
-function App(){
-  return (
-    <MainContainer /> //links to MainContainer.js in navigation folder
-  )
-}
-
-export default App;
