--- Carbon/navigation/screens/index.js
@@ -1,7 +1,0 @@
-export {default as HomeScreen} from './Home/HomeScreen';
-export {default as ProgressScreen} from './Progress/ProgressScreen';
-export {default as ForumScreen} from './Forum/ForumScreen';
-export {default as RankingScreen} from './Ranking/RankingScreen';
-export {default as SettingsScreen} from './Settings/SettingsScreen';
-export {default as QuizScreen} from './Quiz/QuizScreen';
-// export {default as AddProgress} from './Progress/AddProgress';
