+++ src/babel.config.js
