+++ src/containers/ProgressScreen/RecordEmission.js
