+++ src/containers/index.js
@@ -1,0 +1,7 @@
+export {default as HomeScreen} from './HomeScreen';
+export {default as ProgressScreen} from './ProgressScreen';
+export {default as ForumScreen} from './ForumScreen';
+export {default as RankingScreen} from './RankingScreen';
+export {default as SettingsScreen} from './SettingsScreen';
+export {default as QuizScreen} from './QuizScreen';
+// export {default as AddProgress} from './Progress/AddProgress';
