+++ src/package.json
@@ -11,7 +11,9 @@
   },
   "jest": {
     "preset": "jest-expo",
-    "setupFilesAfterEnv": ["@testing-library/jest-native/extend-expect"],
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-native/extend-expect"
+    ],
     "transformIgnorePatterns": [
       "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)"
     ],
@@ -36,7 +38,6 @@
     "expo": "~47.0.12",
     "expo-linear-gradient": "~12.0.1",
     "expo-status-bar": "~1.4.2",
-    "jest": "^26.6.3",
     "jest-expo": "^48.0.2",
     "react": "18.1.0",
     "react-chartjs-2": "^5.2.0",
@@ -56,6 +57,7 @@
     "@babel/core": "^7.12.9",
     "@testing-library/jest-native": "^5.4.2",
     "@testing-library/react-native": "^12.0.1",
+    "jest": "^29.5.0",
     "react-test-renderer": "^18.1.0"
   },
   "private": true
